HCS-10: back - test return error for delete trip if has entries

+++ app/Http/Livewire/Trips/Destroy.php
@@ -12,6 +12,11 @@ class Destroy extends Component
 
     public function save(): void
     {
+        if ($this->trip->entries()->count()) {
+            $this->addError('trip', 'This trip has entries.');
+
+            return;
+        }
 
         $this->trip->delete();
